chore: extract openapi functions into core directory
