+++ src/core/mod.rs
@@ -3,3 +3,4 @@ pub mod command;
 pub mod config;
 pub mod env;
 pub mod ephenv;
+pub mod openapi;
